Created UserController, removed articles and bookmarks services
