--- tests/Unit/Services/FeedbackService/MarkCommentsTest.php
@@ -1,41 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Tests\Unit\Services\FeedbackService;
-
-use App\Models\Article\Article;
-use App\Models\Comment\Comment;
-use App\Models\User\User;
-use App\Service\FeedbackService;
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Tests\TestCase;
-
-class MarkCommentsTest extends TestCase
-{
-    use RefreshDatabase;
-
-    public function testMarkArticles(): void
-    {
-        /** @var User $user */
-        $user = User::factory()->create();
-
-        /** @var Comment $comment01 */
-        $comment01 = Comment::factory()
-            ->forArticle(Article::factory()->create())
-            ->bookmarkedBy($user)
-            ->create();
-
-        /** @var Comment $comment02 */
-        $comment02 = Comment::factory()
-            ->forArticle(Article::factory()->create())
-            ->create();
-
-        $service = new FeedbackService();
-
-        $service->markComments($user, [$comment01, $comment02]);
-
-        $this->assertTrue($comment01->is_bookmarked);
-        $this->assertFalse($comment02->is_bookmarked);
-    }
-}
